+++ validator-string/src/main/java/name/valery1707/jcommander/validators/string/PatternNotMatched.java
@@ -1,0 +1,12 @@
+package name.valery1707.jcommander.validators.string;
+
+import java.util.regex.Pattern;
+
+@SuppressWarnings("WeakerAccess")
+public abstract class PatternNotMatched extends PatternChecker {
+
+	@SuppressWarnings("WeakerAccess")
+	protected PatternNotMatched(Pattern pattern) {
+		super(pattern, true, true);
+	}
+}